refactored ace.system.redis ns
